Fixed imports after refactoring
