+++ cmd/fetch.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/OpenCHAMI/configurator/internal/util"
+	"github.com/OpenCHAMI/configurator/pkg/util"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
