+++ pkg/generator/plugins/hostfile/hostfile.go
@@ -3,9 +3,9 @@ package main
 import (
 	"fmt"
 
-	configurator "github.com/OpenCHAMI/configurator/internal"
-	"github.com/OpenCHAMI/configurator/internal/generator"
-	"github.com/OpenCHAMI/configurator/internal/util"
+	configurator "github.com/OpenCHAMI/configurator/pkg"
+	"github.com/OpenCHAMI/configurator/pkg/generator"
+	"github.com/OpenCHAMI/configurator/pkg/util"
 )
 
 type Hostfile struct{}
